[BugFix] Floating point value could not be used as Max Execute Time of job
Root cause: incorrect property type in the model

+++ DeviceAdministration/Web/Models/ScheduleJobModel.cs
@@ -20,6 +20,6 @@ namespace Microsoft.Azure.Devices.Applications.RemoteMonitoring.DeviceAdmin.Web.
         public DateTime StartDateUtc { get; set; }
 
         [Required]
-        public int MaxExecutionTimeInMinutes { get; set; }
+        public double MaxExecutionTimeInMinutes { get; set; }
     }
 }
